app/requirements.txt
pymongo
flask
plotly_express
flask_wtf
